+++ .gitignore
@@ -74,3 +74,4 @@ com_crashlytics_export_strings.xml
 crashlytics.properties
 crashlytics-build.properties
 
+test_data
