+++ go/operation/compress.go
@@ -1,4 +1,4 @@
-package storage
+package operation
 
 import (
 	"bytes"
